增加了通过excel导入班级和导出榜单excel的功能，为各个app注册了admin网站的管理

+++ schoolclass/views.py
@@ -1,8 +1,14 @@
+import xlrd
+import time
+import random
+import pandas
+
 from django.shortcuts import render, reverse
 from django.views.generic import View, FormView
 from django.http import HttpResponseRedirect
+from django.contrib.auth.hashers import make_password
 
-from .forms import SchoolClassForm, AddStudentForm, UpdateClassForm
+from .forms import SchoolClassForm, AddStudentForm, UpdateClassForm, ExcelInputForm
 from .models import SchoolClass
 from account.models import User
 
@@ -63,8 +69,11 @@ class DeleteClassView(View):
             return HttpResponseRedirect('/reject/')
         if school_class.tercher_id != request.user.id:
             return HttpResponseRedirect('/reject/')
-        SchoolClass.objects.filter(id=class_id).delete()
-        s_url = reverse('schoolclass:show_class', kwargs={'pk': 0})
+        if User.objects.filter(school_class=school_class).exists():
+            s_url = reverse('schoolclass:show_class', kwargs={'pk': class_id})
+        else:
+            SchoolClass.objects.filter(id=class_id).delete()
+            s_url = reverse('schoolclass:show_class', kwargs={'pk': 0})
         return HttpResponseRedirect(s_url)
 
 
@@ -151,8 +160,7 @@ class ShowClassView(FormView):
         else:
             sc = dict()
             sc['id'] = 0
-        return render(request, self.template_name,
-                      {'class_list': school_class, 'student_list': student_list, 'cur_class': sc})
+        return render(request, self.template_name, {'class_list': school_class, 'student_list': student_list, 'cur_class': sc})
 
 
 class ResetPasswordView(View):
@@ -169,3 +177,40 @@ class ResetPasswordView(View):
             return HttpResponseRedirect('/reject/')
         s_url = reverse('schoolclass:show_class', kwargs={'pk': class_id})
         return HttpResponseRedirect(s_url)
+
+
+class AddStudentsByExcelView(View):
+    template_name = 'schoolclass/student_excel.jinja2'
+    form_class = ExcelInputForm
+
+    def get(self, request, *args, **kwargs):
+        if not self.request.user.is_authenticated or not self.request.user.is_superuser:
+            return HttpResponseRedirect('/')
+        form = self.form_class()
+        pk = kwargs.get('pk')
+        return render(request, self.template_name, {'form': form, 'schoolclass_id': pk})
+
+    def post(self, request, *args, **kwargs):
+        form = self.form_class(request.POST, request.FILES)
+        if form.is_valid():
+            pk = kwargs.get('pk')
+            # data = xlrd.open_workbook(filename=None, file_contents=self.request.FILES['excel'].read())
+            # sheet = data.sheet_by_index(0)
+            # n = sheet.nrows
+            # username = str(sheet.cell_value(i, 0))
+            df = pandas.read_excel(self.request.FILES['excel'].read())
+            data = df.values.tolist()
+            n = len(data)
+            for i in range(n):
+                username = str(data[i][0])
+                name = str(data[i][1])
+                password = make_password(str(data[i][2]))
+                print(password)
+                if User.objects.filter(username=username).exists():
+                    User.objects.filter(username=username).update(name=name, password=password, school_class_id=pk)
+                    # user.save(update_fields=['name', 'password', 'school_class_id'])
+                else:
+                    ticks = int(time.time())
+                    fake_email = str(ticks) + str(random.randint(1, 100)) + str(random.randint(1, 100)) + '@' + 'qq.com'
+                    User.objects.create(username=username, password=password, email=fake_email, name=name, school_class_id=pk)
+        return HttpResponseRedirect(reverse('schoolclass:show_class', kwargs={'pk': pk}))
